% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ramet_sets.r
\name{collect_ramet_sets}
\alias{collect_ramet_sets}
\title{Find sets of ramets, based on a distance matrix 
and a nominated threshold}
\usage{
collect_ramet_sets(dm, dthresh, lt = TRUE)
}
\arguments{
\item{dm}{- distance matrix for a set of individuals [Required]}

\item{dthresh}{- threshold value for corresponding measure in matrix [Required]}
}
\value{
a vector with assignments to ramet sets
}
\description{
Find sets of ramets, based on a distance matrix 
and a nominated threshold
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

