% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_progeny.r
\name{count_progeny}
\alias{count_progeny}
\title{Calculate genotype counts for all pairs of individuals}
\usage{
count_progeny(pop, fam, tissue)
}
\arguments{
\item{fam}{- vector assigning individuals to a family [Required]}

\item{tissue}{- vector indicating if mother (M) or progeny (P) [Required]}

\item{gt}{- genotype matrix for a set of individuals (in rows) [Required]}
}
\value{
a list for each pair of individuals
}
\description{
Calculate genotype counts for all pairs of individuals
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

