% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_selected_ramets.r
\name{prune_selected_ramets}
\alias{prune_selected_ramets}
\title{Remove chosen ramets}
\usage{
prune_selected_ramets(ps, gt)
}
\arguments{
\item{ps}{- a list of samples to prune [Required]}

\item{gt}{- genotype matrix, with missing data as NA}
}
\value{
a list consisting of two matrices
}
\description{
Remove chosen ramets
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

