% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramet_sets_from_columns.r
\name{ramet_sets_from_columns}
\alias{ramet_sets_from_columns}
\title{Encode known ramet sets using a meta data file}
\usage{
ramet_sets_from_columns(ramet_genet)
}
\arguments{
\item{ramet_genet}{- two columns: names of samples and genets [Required]}
}
\value{
a list consisting of two matrices
}
\description{
Encode known ramet sets using a meta data file
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

