% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramet_summary.r
\name{ramet_summary}
\alias{ramet_summary}
\title{Summarize measure values for ramets and non-ramets}
\usage{
ramet_summary(ramet_set, measure_matrix)
}
\arguments{
\item{ramet_set}{- a list of samples to prune [Required]}

\item{measure_matrix}{- genotype matrix, with missing data as NA}
}
\value{
a list
}
\description{
Summarize measure values for ramets and non-ramets
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

