% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramets_to_prune.r
\name{ramets_to_prune}
\alias{ramets_to_prune}
\title{Generate a list of ramets to prune, either randomly, 
or based on missingness}
\usage{
ramets_to_prune(rs, method = "random", gt = NULL)
}
\arguments{
\item{rs}{- a ramets set, as generated by collect_ramet_sets [Required]}

\item{method}{- either "random" or "least_missing"}

\item{gt}{- genotype matrix, with missing data as NA}
}
\value{
a list consisting of two matrices
}
\description{
Generate a list of ramets to prune, either randomly, 
or based on missingness
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

