% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_progeny.r
\name{resample_progeny}
\alias{resample_progeny}
\title{Calculate genotype counts for all pairs of individuals}
\usage{
resample_progeny(pop, fam, tissue, nseed, nfam, npop = NULL)
}
\arguments{
\item{fam}{- vector assigning individuals to a family [Required]}

\item{tissue}{- vector indicating if mother (M) or progeny (P) [Required]}
}
\value{
a list for each pair of individuals
}
\description{
Calculate genotype counts for all pairs of individuals
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

