% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_het_measures.r
\name{shared_het_measures}
\alias{shared_het_measures}
\title{Calculate genotype counts for all pairs of individuals}
\usage{
shared_het_measures(gt)
}
\arguments{
\item{gt}{- genotype matrix for a set of individuals (in rows) [Required]}
}
\value{
a list for each pair of individuals
}
\description{
Calculate genotype counts for all pairs of individuals
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

