% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_dataset.r
\name{simulated_dataset}
\alias{simulated_dataset}
\title{make a dataset with known ramets}
\usage{
simulated_dataset(gt, neach = 1, het2hom = 0.01, hom2het = 0.001,
  hom2hom = 0.001, simulate_source = TRUE)
}
\arguments{
\item{gt}{- a genotype matrix 0, 1, 2 format [Required]}

\item{neach}{- number of ramets of each sample}

\item{het2hom}{- mutation rate het to hom}

\item{het2hom}{- mutation rate hom to het}

\item{het2hom}{- mutation rate hom to hom}
}
\value{
a list consisting of two matrices
}
\description{
make a dataset with known ramets
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

