% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_ramet.r
\name{simulated_ramet}
\alias{simulated_ramet}
\title{make a ramet with simulated sequencing errors}
\usage{
simulated_ramet(gv, het2hom = 0.01, hom2het = 0.001, hom2hom = 0.001)
}
\arguments{
\item{gv}{- a vector of genotypes, in 0, 1, 2 format [Required]}

\item{het2hom}{- mutation rate het to hom}

\item{het2hom}{- mutation rate hom to het}

\item{het2hom}{- mutation rate hom to hom}
}
\value{
a list consisting of two matrices
}
\description{
make a ramet with simulated sequencing errors
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

