% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_all_pairs.r
\name{table_all_pairs}
\alias{table_all_pairs}
\title{Calculate genotype counts for all pairs of individuals}
\usage{
table_all_pairs(gt)
}
\arguments{
\item{gt}{- genotype matrix for a set of individuals (in rows) [Required]}
}
\value{
a list for each pair of individuals
}
\description{
Calculate genotype counts for all pairs of individuals
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

