% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_genotypes.r
\name{tabulate_genotypes}
\alias{tabulate_genotypes}
\title{For a pair of individuals, count shared genotypic states}
\usage{
tabulate_genotypes(gtp)
}
\arguments{
\item{gtp}{- genotype matrix for two individuals (in rows) [Required]}
}
\value{
a list consisting of two matrices
}
\description{
For a pair of individuals, count shared genotypic states
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}

